extract HorizontalScrollArrow component
