+++ src/components/HorizontalScroll/HorizontalScrollArrow.tsx
@@ -1,0 +1,20 @@
+import React, { FC } from 'react';
+import Icon24Chevron from '@vkontakte/icons/dist/24/chevron_right';
+
+export interface HorizontalScrollArrowProps {
+  onClick: () => void;
+  direction: 'left' | 'right';
+}
+
+const HorizontalScrollArrow: FC<HorizontalScrollArrowProps> = (props) => {
+  const { onClick, direction } = props;
+  return (
+    <div className={`HorizontalScroll__arrow HorizontalScroll__arrow-${direction}`} onClick={onClick}>
+      <div className="HorizontalScroll__arrow-icon">
+        <Icon24Chevron />
+      </div>
+    </div>
+  );
+};
+
+export default HorizontalScrollArrow;
